tests/wpt/tests/html/canvas/element/text/2d.text.measure.width.space.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.text.measure.width.space</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<style>
@font-face {
  font-family: CanvasTest;
  src: url("/fonts/CanvasTest.ttf");
}
</style>
<body class="show_output">

<h1>2d.text.measure.width.space</h1>
<p class="desc">Space characters are converted to U+0020 and NOT collapsed</p>


<span style="font-family: CanvasTest; position: absolute; visibility: hidden">A</span>
<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
promise_test(async t => {

  var canvas = document.getElementById('c');
  var ctx = canvas.getContext('2d');

  await document.fonts.ready;
  ctx.font = '50px CanvasTest';
  _assertSame(ctx.measureText('A B').width, 150, "ctx.measureText('A B').width", "150");
  _assertSame(ctx.measureText('A  B').width, 200, "ctx.measureText('A  B').width", "200");
  _assertSame(ctx.measureText('A \x09\x0a\x0c\x0d  \x09\x0a\x0c\x0dB').width, 650, "ctx.measureText('A \\x09\\x0a\\x0c\\x0d  \\x09\\x0a\\x0c\\x0dB').width", "650");
  _assert(ctx.measureText('A \x0b B').width >= 200, "ctx.measureText('A \\x0b B').width >= 200");

  _assertSame(ctx.measureText(' AB').width, 150, "ctx.measureText(' AB').width", "150");
  _assertSame(ctx.measureText('AB ').width, 150, "ctx.measureText('AB ').width", "150");

}, "Space characters are converted to U+0020 and NOT collapsed");
</script>
